StructHelpers

Documentation for StructHelpers.

StructHelpers.has_batteries

StructHelpers.hash_eq_as

StructHelpers.@batteries

StructHelpers.@enumbatteries

StructHelpers.has_batteries — Method

has_batteries(T::Type)::Bool

Check if @batteries or @enumbatteries was applied to T.

StructHelpers.hash_eq_as — Method

hash_eq_as(obj)

This allows to fine tune the behavior or hash, == and isequal for structs decorated by @batteries. For instances the generated isequal method looks like this:

function Base.isequal(o1::T, o2::T)
    proxy1 = StructHelpers.hash_eq_as(o1)
    proxy2 = StructHelpers.hash_eq_as(o2)
    isequal(proxy1, proxy2)
end

Overloading hash_eq_as is useful for instance if you want to skip certain fields of obj or handle them in a special way.

StructHelpers.@batteries — Macro

@batteries T [options]

Automatically derive several methods for type T.

Example

struct S
    a
    b
end

@batteries S
@batteries S hash=false # don't overload `Base.hash`
@batteries S kwconstructor=true # add a keyword constructor

Supported options and defaults are:

eq = true:

Define Base.(==) structurally.

isequal = true:

Define Base.isequal structurally.

hash = true:

Define Base.hash structurally.

kwconstructor = false:

Add a keyword constructor.

selfconstructor = true:

Add a constructor of the for T(self::T) = self

kwshow = false:

Overload Base.show such that the names of each field are printed.

getproperties = true:

Overload ConstructionBase.getproperties.

constructorof = true:

Overload ConstructionBase.constructorof.

typesalt = nothing:

Only used if hash=true. In this case the hash will be purely computed from typesalt and hash_eq_as(obj). The type T will not be used otherwise. This makes the hash more likely to stay constant, when executing on a different machine or julia version

StructTypes = false:

Overload StructTypes.StructType to be Struct(). Needs the StructTypes.jl package to be installed.

See also hash_eq_as

StructHelpers.@enumbatteries — Macro

@enumbatteries T [options]

Automatically derive several methods for Enum type T.

Example

@enum Color Red Blue Yellow
@enumbatteries Color
@enumbatteries Color hash=false # don't overload `Base.hash`
@enumbatteries Color symbol_conversion=true # allow convert(Color, :Blue), Color(:Blue), convert(Symbol, Blue), Symbol(Blue)

Supported options and defaults are:

string_conversion = false:

Add convert(MyEnum, ::String), MyEnum(::String), convert(String, ::MyEnum) and String(::MyEnum)

symbol_conversion = false:

Add convert(MyEnum, ::Symbol), MyEnum(::Symbol), convert(Symbol, ::MyEnum) and Symbol(::MyEnum)

selfconstructor = true:

Add a constructor of the for T(self::T) = self